fix: add back execute_metered

+++ crates/vm/src/arch/vm.rs
@@ -475,14 +475,14 @@ where
             let ctx = MeteredCtx::new(
                 widths.len(),
                 continuations_enabled,
-                num_access_adapters,
+                num_access_adapters as u8,
                 segment
                     .chip_complex
                     .memory_controller()
                     .memory
                     .min_block_size
                     .iter()
-                    .map(|&x| log2_strict_usize(x as usize))
+                    .map(|&x| log2_strict_usize(x as usize) as u8)
                     .collect(),
                 segment
                     .chip_complex
